Move data structures into their own folder
