+++ tools/chall-parser/parser/compose/__init__.py
@@ -1,0 +1,18 @@
+from parser.compose.compose import ComposeFile, Network, ComposeResourceName, ServicesDict, NetworksDict
+from parser.compose.service import Service, Template
+from parser.compose.challenge_info import ChallengeInfo, TextHint, Hint, Question, Variable
+
+__all__ = [
+    'ComposeFile',
+    'Service',
+    'Template',
+    'ChallengeInfo',
+    'TextHint',
+    'Hint',
+    'Question',
+    'Variable',
+    'Network',
+    'ComposeResourceName',
+    'ServicesDict',
+    'NetworksDict'
+]
