+++ tools/chall-parser/parser/compose/service.py
